all: fix crash caused by paths containing '//' on linux #97 and #105 mesh: improve speed weighting

+++ libs/Common/Types.h
@@ -234,7 +234,7 @@ typedef signed int			INT;
 typedef unsigned int		UINT;
 typedef long				LONG;
 
-typedef LONG				HRESULT;
+typedef int32_t				HRESULT;
 
 typedef CHAR*				LPSTR;
 typedef const CHAR*			LPCSTR;
